wx2.9x64 was tripping over some format calls with "%1$d" and size_t. Since all string items are 'd', change the format to cast long/size_t to int. In addition, entry-validation was triggering when selecting a page in wxNotebook - use ChangeSelection instead of SetSelection

+++ src/Win/DlgTrial.cpp
@@ -559,7 +559,8 @@ void CDlgTrial::OnOk(wxCommandEvent& evt)
 			}
 			if (0 < nDelete)
 			{
-				wxString msg = wxString::Format(_("IDS_CONFIG_DELETE_RUNS"), m_pTrial->GetRuns().size());
+				wxString msg = wxString::Format(_("IDS_CONFIG_DELETE_RUNS"),
+					static_cast<int>(m_pTrial->GetRuns().size()));
 				if (wxYES != wxMessageBox(msg, wxMessageBoxCaptionStr, wxYES_NO | wxNO_DEFAULT | wxCENTRE | wxICON_WARNING))
 					return;
 				m_pTrial->GetRuns().clear();
